JavaScriptによるいいね機能の実装
