+++ app/controllers/likes_controller.rb
@@ -10,14 +10,16 @@ class LikesController < ApplicationController
   def create
     article = Article.find(params[:article_id])
     article.likes.create!(user_id: current_user.id)
-    redirect_to article_path(article)
+
+    render json: { status: 'ok' }
   end
 
   def destroy
     article = Article.find(params[:article_id])
     like = article.likes.find_by!(user_id: current_user.id)
     like.destroy!
-    redirect_to article_path(article)
+
+    render json: { status: 'ok' }
   end
 
 end
