+++ app/javascript/packs/application.js
@@ -18,25 +18,3 @@ require("channels")
 
 require("trix")
 require("@rails/actiontext")
-
-import $ from 'jquery'
-import axios from 'axios'
-
-const handleHeartDisplay = (hasLiked) => {
-  if (hasLiked) {
-    $('.active-heart').removeClass('hidden')
-  } else {
-    $('.inactive-heart').removeClass('hidden')
-  }
-}
-
-document.addEventListener('turbolinks:load', () => {
-  const dataset = $('#article-show').data()
-  const articleId = dataset.articleId
-
-  axios.get(`/articles/${articleId}/like`)
-    .then((response) => {
-      const hasLiked = response.data.hasLiked
-      handleHeartDisplay(hasLiked)
-    })
-})
