+++ app/views/articles/show.html.haml
@@ -21,11 +21,9 @@
     = @article.content
   - if user_signed_in?
     .article_heart.hidden.active-heart
-      = link_to article_like_path(@article), data: { method: 'delete' } do
-        = image_tag 'heart-active.svg'
+      = image_tag 'heart-active.svg'
     .article_heart.hidden.inactive-heart
-      = link_to article_like_path(@article), data: { method: 'post' } do
-        = image_tag 'heart.svg'
+      = image_tag 'heart.svg'
 
 .article
   %h2 コメント一覧
@@ -37,3 +35,5 @@
   =link_to new_article_comment_path(@article) do
     .btn-secondary
       コメントを追加
+
+= javascript_pack_tag 'article'
