+++ package.json
@@ -9,6 +9,7 @@
     "@rails/webpacker": "5.4.4",
     "axios": "1.6.8",
     "jquery": "^3.7.1",
+    "rails-ujs": "^5.2.8-1",
     "trix": "^1.2.0",
     "turbolinks": "^5.2.0",
     "webpack": "^4.46.0",
